requirements.txt
GitPython==0.3.2.RC1
async==0.6.1
boto==2.38.0
gitdb==0.5.4
smmap==0.8.2
wsgiref==0.1.2
